Jetpack: Move comment notification override back to the constructor (#15461)
The code block that overrides comment notifications and points the notification
links to Calypso was recently moved to Jetpack::configure(). It was moved so that it
would be called during the plugins_loaded action, which is required by the current
autoloader. Unfortunately, this broke the override because this code block overrides
pluggable functions, and the pluggable functions are loaded just before plugins_loaded fires.

To fix this, move the comment notification override code block back to the Jetpack constructor.
Also, add the Jetpack_Options class to the Autoloader's ignore list because this class is
loaded before plugins_loaded.

Committed via a GitHub action: https://github.com/automattic/jetpack/runs/85337653