Autoloader: remove the defined('JETPACK_AUTOLOAD_DEV') checks from the tests (#17417)
A few of the unit tests require that the JETPACK_AUTOLOAD_DEV constant be set
to true. Currently, we're checking whether the constant is defined before setting
it. If the constant is already defined, we can't guarantee that the constant is
set to true, making the test less reliable.

Let's remove the defined() checks. If the constant is already defined when the unit
tests are executed, an error should be generated.

Committed via a GitHub action: https://github.com/automattic/jetpack/runs/296158023