Autoloader: Reuse an existing autoloader suffix if available (#17591)
The change of suffix can cause transient problems with deployment if a
request comes in while the files are being overwritten. To help mitigate
this, reuse the existing suffix if vendor/autoload_packages.php exists.

And since we're already using Composer's own "autoloader-suffix" config
setting when it's set, we may as well also reuse a suffix from
vendor/autoload.php if we can detect it.

Committed via a GitHub action: https://github.com/automattic/jetpack/runs/337953470