Autoloader: test suite refactor (#18538)
This refactor adjusts the test suite so that it will run against actual installations of the autoloader instead of the `src` directory. By testing the autoloader in the environment it will exist in we can provide a greater degree of confidence and enable better testing of autoloader conflicts.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/552210829