Autoloader: Added Code Coverage Transformation (#19012)
Since the code coverage for the package is executed against
generated autoloaders, it becomes necessary for us to transform
the output to merge the data. This solution employs the use of the
raw PHP report type along with dedicated handlers that can take
the output of this report and generate a Clover report from it.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/625177447