autoloader: Preload classes on plugin upgrade to avoid mid-upgrade fatals (#23039)
It's possible that the new version of the plugin will move some files
around that some later hook will try to load, causing a fatal as the
autoloader will still be looking in the old path. To avoid this, have
the Jetpack autoloader detect when an upgrade is happening and pre-load
all the classes in the plugin.

We do the same for plugin removal, for similar reasons.

Note this doesn't work for classes loaded via PSR-0 or PSR-4. But that
shouldn't matter for us since we always generate an optimized (classmap)
autoloader for releases of our plugins.

Also of note is that this won't work if you're manually installing a
specific version of the plugin as `wp plugin install jetpack --version=10.5`,
as wp-cli does that as an "install" rather than an "upgrade" operation.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/1923832278